[AI] After review fixes

+++ default/python/AI/CombatRatingsAI.py
@@ -60,8 +60,8 @@ class ShipCombatStats(object):
             :type shields: int|None
             :return:
             """
-            self.structure = 1 if structure is None else structure
-            self.shields = 0 if shields is None else shields
+            self.structure = 1.0 if structure is None else structure
+            self.shields = 0.0 if shields is None else shields
             self.attacks = {} if attacks is None else tuple_to_dict(attacks)  # type: dict[float, int]
 
         def get_stats(self, hashable=False):
@@ -149,7 +149,7 @@ class ShipCombatStats(object):
                         # TODO: Depending on future implementation, might actually need to handle this case.
                         warn("Multiple hangar types present on one ship, estimates expected to be wrong.")
                     fighter_damage = max(fighter_damage, part_damage)
-        self._basic_stats = self.BasicStats(attacks, int(structure), int(shields))
+        self._basic_stats = self.BasicStats(attacks, structure, shields)
         self._fighter_stats = self.FighterStats(fighter_capacity, fighter_launch_rate, fighter_damage)
 
     def get_basic_stats(self, hashable=False):
@@ -157,7 +157,7 @@ class ShipCombatStats(object):
 
         :param hashable: if true, returns tuple instead of attacks-dict
         :return: attacks, structure, shields
-        :rtype: (dict|tuple, int, int)
+        :rtype: (dict|tuple, float, float)
         """
         return self._basic_stats.get_stats(hashable=hashable)
